+++ openapi.yaml
@@ -118,7 +118,7 @@ components:
           type: boolean
           default: false
 paths:
-  /:
+  /v1/:
     get:
       summary: Retrieve available Layers
       responses:
@@ -132,7 +132,7 @@ paths:
                   $ref: '#/components/schemas/Layer'
         204:
           description: No layers available
-  /{layer-ref}/:
+  /v1/{layer-ref}/:
     parameters:
       - $ref: '#/components/parameters/LayerID'
     get:
@@ -152,7 +152,7 @@ paths:
                 $ref: '#/components/schemas/Layer'
 
 
-  /content/{layer-ref}/:
+  /v1/content/{layer-ref}/:
     parameters:
       - $ref: '#/components/parameters/LayerID'
     get:
@@ -175,7 +175,7 @@ paths:
                   information is available on a single entry of this response.
                 items:
                   $ref: "#/components/schemas/Object"
-  /content/{layer-ref}/filtered:
+  /v1/content/{layer-ref}/filtered:
     parameters:
       - $ref: '#/components/parameters/LayerID'
     get:
