+++ routes/v1/filtered-objects.go
